moved item function from GameController to ItemController
