+++ lib/controllers/GameController.js
@@ -20,11 +20,11 @@ class GameController {
     // JET Basics //
     static async modeOfflinePatches(_request = null, reply = null) {
         return FastifyResponse.zlibJsonReply(reply, database.core.serverConfig.Patches);
-    };
+    }
 
     static async modeOfflinePatchNodes(_request = null, reply = null) {
         return FastifyResponse.zlibJsonReply(reply, database.core.serverConfig.PatchNodes);
-    };
+    }
 
     // Game //
 
@@ -39,7 +39,7 @@ class GameController {
                 reply,
                 FastifyResponse.applyBody({ utc_time: await getCurrentTimestamp() }, 999, "Profile Not Found!!"));
         }
-    };
+    }
 
     static async clientGameVersionValidate(request = null, reply = null) {
         logger.logInfo("Client connected with version: " + request.body.version.major);
@@ -48,7 +48,7 @@ class GameController {
                 reply,
                 FastifyResponse.applyBody(null)
             );
-    };
+    }
 
     static async clientGameConfig(request = null, reply = null) {
         const sessionID = await FastifyResponse.getSessionID(request);
@@ -76,7 +76,7 @@ class GameController {
                 reply,
                 FastifyResponse.applyBody(responseObject)
             );
-    };
+    }
 
     static async clientGameKeepAlive(request = null, reply = null) {
         let msg = "OK";
@@ -95,7 +95,7 @@ class GameController {
             reply,
             FastifyResponse.applyBody({ msg: msg, utc_time: await getCurrentTimestamp() })
         );
-    };
+    }
 
     static async clientProfileList(request = null, reply = null) {
         const output = [];
@@ -121,7 +121,7 @@ class GameController {
             reply,
             FastifyResponse.applyBody(output)
         );
-    };
+    }
 
     static async clientProfileSelect(request, reply) {
         const sessionID = await FastifyResponse.getSessionID(request);
@@ -134,14 +134,14 @@ class GameController {
             reply,
             FastifyResponse.applyBody(output)
         );
-    };
+    }
 
     static async clientGameProfileNicknameReserved(_request = null, reply = null) {
         return FastifyResponse.zlibJsonReply(
             reply,
             FastifyResponse.applyBody("")
         );
-    };
+    }
 
     static async clientGameProfileNicknameValidate(request = null, reply = null) {
         const validate = await Profile.ifAvailableNickname(request.body.nickname);
@@ -163,7 +163,7 @@ class GameController {
                     FastifyResponse.applyBody({ status: "ok" })
                 );
         }
-    };
+    }
 
     static async clientGameProfileCreate(request = null, reply = null) {
         const sessionID = await FastifyResponse.getSessionID(request);
@@ -213,7 +213,7 @@ class GameController {
             reply,
             FastifyResponse.applyBody({ uid: "pmc" + sessionID })
         );
-    };
+    }
 
     static async clientGameProfileVoiceChange(request = null, reply = null) {
         const playerProfile = await Profile.get(await FastifyResponse.getSessionID(request));
@@ -223,10 +223,10 @@ class GameController {
             reply,
             FastifyResponse.applyBody({
                 status: 0,
-                nicknamechangedate: ~~(new Date() / 1000)
+                nicknamechangedate: ~~(await getCurrentTimestamp())
             })
         );
-    };
+    }
 
     static async clientGameProfileNicknameChange(request = null, reply = null) {
         const playerProfile = await Profile.get(await FastifyResponse.getSessionID(request));
@@ -234,28 +234,16 @@ class GameController {
 
         switch (validate) {
             case "tooshort":
-                return FastifyResponse.zlibJsonReply(
-                    reply,
-                    FastifyResponse.applyBody(null, 256, "256 -")
-                );
+                return FastifyResponse.zlibJsonReply(reply, FastifyResponse.applyBody(null, 256, "256 -"));
             case "taken":
-                return FastifyResponse.zlibJsonReply(
-                    reply,
-                    FastifyResponse.applyBody(null, 255, "255 - ")
-                );
+                return FastifyResponse.zlibJsonReply(reply, FastifyResponse.applyBody(null, 255, "255 - "));
             default:
                 playerProfile.character.Info.Nickname = request.body.nickname;
                 playerProfile.character.Info.LowerNickname = request.body.nickname.toLowerCase();
                 await playerProfile.saveCharacter();
-                return FastifyResponse.zlibJsonReply(
-                    reply,
-                    FastifyResponse.applyBody({
-                        status: 0,
-                        nicknamechangedate: await getCurrentTimestamp()
-                    })
-                );
+                return FastifyResponse.zlibJsonReply( reply, FastifyResponse.applyBody({status: 0,nicknamechangedate: await getCurrentTimestamp()}));
         }
-    };
+    }
 
     static async clientGameProfileAcceptQuest(moveAction = null, _reply = null, playerProfile = null) {
         const quest = await Quest.get(moveAction.qid);
@@ -272,85 +260,7 @@ class GameController {
         };
         await playerProfile.addDialogue(quest.traderId, messageContent, questReward);
         return {};
-    };
-
-    static async clientGameProfileMoveItem(moveAction = null, _reply = null, playerProfile = null) {
-        logger.logDebug("Move request:");
-        logger.logDebug(moveAction);
-
-        const movedItems = await playerProfile.character.moveItems(moveAction);
-        if (!movedItems) {
-            logger.logError("[clientGameProfileMoveItem] Failed to move items, report to a developer.");
-        }
-        return {};
     }
-
-    static async clientGameProfileExamine(moveAction = null, reply = null, playerProfile = null) {
-        logger.logDebug("Examine request:");
-        logger.logDebug(moveAction);
-
-        let templateItem;
-
-        if (typeof moveAction.fromOwner !== "undefined") {
-            switch (moveAction.fromOwner.type) {
-                case "Trader":
-                    const trader = await Trader.get(moveAction.fromOwner.id);
-                    if (trader) {
-                        const inventoryItem = await trader.getAssortItemByID(moveAction.item);
-                        if (inventoryItem) {
-                            templateItem = await Item.get(inventoryItem._tpl);
-                        } else {
-                            logger.logError(`[clientGameProfileExamine] Examine Request failed: Unable to find item database template of itemId ${moveAction.item}`);
-                            return false;
-                        }
-                    } else {
-                        logger.logError("[clientGameProfileExamine] Examine Request failed: Unable to get trader data.");
-                        return false;
-                    }
-                    break;
-
-                case "RagFair":
-                    const ragfair = await Ragfair.get("FleaMarket");
-                    const ragfairOffers = ragfair.offers;
-                    const item = ragfairOffers.find(function (i) {
-                        if (i._id === moveAction.fromOwner.id) return i;
-                    });
-                    templateItem = await Item.get(item.items[0]._tpl);
-                    break;
-
-                case "HideoutUpgrade":
-                case "HideoutProduction":
-                case "ScavCase":
-                    templateItem = await Item.get(moveAction.item);
-                    break;
-
-                default:
-                    logger.logError(`[clientGameProfileExamine] Examine Request failed: Unknown moveAction.fromOwner.Type: ${moveAction.fromOwner.type}`);
-                    return false;
-            }
-        } else {
-            const item = await playerProfile.character.getInventoryItemByID(moveAction.item);
-            if (item) {
-                templateItem = await Item.get(item._tpl);
-            } else {
-                logger.logError(`[clientGameProfileExamine] Examine Request failed: Unable to find item database template of itemId ${moveAction.item}`);
-                return false;
-            }
-        }
-
-        if (templateItem) {
-            if (await playerProfile.character.examineItem(templateItem._id)) {
-                await playerProfile.character.addExperience(templateItem._props.ExamineExperience);
-            } else {
-                logger.logError(`[clientGameProfileExamine] Examine Request failed: Unable to examine itemId ${templateItem._id}`);
-            }
-        } else {
-            // this will crash because inventoryItem can't be reached
-            logger.logError(`[clientGameProfileExamine] Examine Request failed: Unable to find item database template of itemId ${inventoryItem._tpl}`);
-        }
-
-        return {};
-    };
 
     static async clientGameProfileTradingConfirm(moveAction = null, _reply = null, playerProfile = null) {
         logger.logDebug("[clientGameProfileTradingConfirm] Trading request:");
@@ -368,7 +278,6 @@ class GameController {
             const traderAssort = await trader.getFilteredAssort(playerProfile);
             const traderItemChildren = await traderItem.getAllChildItemsInInventory(traderAssort.items);
             const traderItemTemplate = await Item.get(traderItem._tpl);
-
 
             let preparedChildren = false;
             if (traderItemChildren) {
@@ -475,7 +384,6 @@ class GameController {
             const offer = await RagfairOffer.getById(moveAction.offers[0].id, ragfair.offers);
             const itemTemplate = await Item.get(offer.items[0]._tpl);
 
-            //const unpreparedChildren = await Item.getAllChildItemsInInventory(item.items);
             let preparedChildren = false;
             if (offer.items.length > 0) {
                 preparedChildren = await Item.prepareChildrenForAddItem(offer.items[0], offer.items);
@@ -534,9 +442,10 @@ class GameController {
                             await trader.removeItemFromAssortAfterBuy({item_id: offer.root, count: moveAction.offers[0].count});
                         }
                     }
-                    /*await trader.reduceStock(requestEntry.item_id, requestEntry.count);*/
                 }
-            } else { logger.logError(`[clientGameProfileTradingConfirm] Unable to add items`); }
+            } else {
+                logger.logError(`[clientGameProfileTradingConfirm] Unable to add items`);
+            }
 
             logger.logDebug(output);
             logger.logDebug(output.items);
@@ -545,97 +454,7 @@ class GameController {
             logger.logError(`[clientGameProfileTradingConfirm] My brother in christ what are you trying to do ? ${moveAction.type} ? That shit is not done lmao pay me now.`);
         }
         return output;
-    };
-
-    static async clientGameProfileSplitItem(moveAction = null, _reply = null, playerProfile = null) {
-        const splittedItems = await playerProfile.character.splitItems(moveAction);
-        if (splittedItems) {
-            return {
-                items: { new: [splittedItems] }
-            };
-        }
-    };
-
-    static async clientGameProfileMergeItem(moveAction = null, _reply = null, playerProfile = null) {
-        const mergedItems = await playerProfile.character.mergeItems(moveAction);
-        if (mergedItems) {
-            return {
-                items: { del: [mergedItems] }
-            };
-        }
-    };
-
-    static async clientGameProfileRemoveItem(moveAction = null, _reply = null, playerProfile = null) {
-        const deletedItems = await playerProfile.character.removeItems(moveAction);
-        if (deletedItems) {
-            return {
-                items: { del: [deletedItems] }
-            };
-        }
-    };
-
-    static async clientGameProfileFoldItem(moveAction = null, _reply = null, playerProfile = null) {
-        if (playerProfile) {
-            const item = await playerProfile.character.getInventoryItemByID(moveAction.item);
-            if (item) {
-                if (typeof item.upd === "undefined") {
-                    item.upd = {};
-                }
-
-                if (typeof item.upd.Foldable === "undefined") {
-                    item.upd.Foldable = {};
-                }
-
-                item.upd.Foldable.Folded = moveAction.value;
-            }
-        }
-    };
-
-    static async clientGameProfileTagItem(moveAction = null, _reply = null, playerProfile = null) {
-        if (playerProfile) {
-            const item = await playerProfile.character.getInventoryItemByID(moveAction.item);
-            if (item) {
-                if (typeof item.upd === "undefined") {
-                    item.upd = {};
-                }
-
-                if (typeof item.upd.Tag === "undefined") {
-                    item.upd.Tag = {};
-                }
-
-                item.upd.Tag.Color = moveAction.TagColor;
-                item.upd.Tag.Name = moveAction.TagName;
-            }
-        }
-    };
-
-    static async clientGameProfileToggleItem(moveAction = null, _reply = null, playerProfile = null) {
-        if (playerProfile) {
-            const item = await playerProfile.character.getInventoryItemByID(moveAction.item);
-            if (item) {
-                if (typeof item.upd === "undefined") {
-                    item.upd = {};
-                }
-
-                if (typeof item.upd.Togglable === "undefined") {
-                    item.upd.Togglable = {};
-                }
-
-                item.upd.Togglable.On = moveAction.value;
-            }
-        }
-    };
-
-    static async clientGameProfileBindItem(moveAction = null, _reply = null, playerProfile = null) {
-        if (playerProfile) {
-            for (const index in playerProfile.character.Inventory.fastPanel) {
-                if (playerProfile.character.Inventory.fastPanel[index] === moveAction.item) {
-                    playerProfile.character.Inventory.fastPanel[index] = "";
-                }
-            }
-            playerProfile.character.Inventory.fastPanel[moveAction.index] = moveAction.item;
-        }
-    };
+    }
 
     static async clientGameProfileReadEncyclopedia(moveAction = null, _reply = null, playerProfile = null) {
         if (playerProfile) {
@@ -643,7 +462,7 @@ class GameController {
                 playerProfile.character.Encyclopedia[id] = true;
             }
         }
-    };
+    }
 
     static async clientGameProfileHideoutUpgrade(moveAction = null, _reply = null, playerProfile = null) {
         logger.logDebug(moveAction);
@@ -882,7 +701,7 @@ class GameController {
                 return;
             }
         }
-    };
+    }
 
     static async clientGameProfileHideoutContinuousProductionStart(moveAction = null, _reply = null, playerProfile = null) {
         if (playerProfile) {
@@ -922,7 +741,7 @@ class GameController {
             );
             await playerPMC.save();
         }
-    };
+    }
 
     static async clientGameProfileEditNote(moveAction = null, _reply = null, playerProfile = null) {
         if (playerProfile) {
@@ -934,7 +753,7 @@ class GameController {
             };
             await playerPMC.save();
         }
-    };
+    }
 
     static async clientGameProfileRemoveNote(moveAction = null, _reply = null, playerProfile = null) {
         if (playerProfile) {
@@ -943,7 +762,7 @@ class GameController {
             delete playerPMC.Notes.Notes[moveAction.index];
             await playerPMC.save();
         }
-    };
+    }
 
     static async clientGameProfileResetWishList(moveAction = null, _reply = null, playerProfile = null) {
         if (playerProfile) {
@@ -952,7 +771,7 @@ class GameController {
             playerPMC.WishList = [];
             await playerPMC.save();
         }
-    };
+    }
 
     static async clientGameProfileHideoutScavCaseProductionStart(moveAction = null, _reply = null, playerProfile = null) {
         const output = {
@@ -971,21 +790,6 @@ class GameController {
 
             const products = await hideoutScavcaseProduction.generateRewards();
 
-            /**
-             * M4A1 TEST
-             */
-            //products.push({
-            //    _id: await generateMongoID(),
-            //    _tpl: "5447a9cd4bdc2dbd208b4567"
-            //});
-
-            /**
-             * Casque TEST
-             */
-            //products.push({
-            //    _id: await generateMongoID(),
-            //    _tpl: "5b432b965acfc47a8774094e"
-            //});
             if (itemTaken) {
                 output.items.change = itemTaken.changed;
                 output.items.removed = itemTaken.removed;
@@ -1003,7 +807,7 @@ class GameController {
             }
         }
         return output;
-    };
+    }
 
     static async clientGameProfileHideoutTakeProduction(moveAction = null, _reply = null, playerProfile = null) {
         const output = {
@@ -1042,7 +846,7 @@ class GameController {
             await playerProfile.character.removeHideoutProductionById(moveAction.recipeId);
         }
         return output;
-    };
+    }
 
     static async clientGameProfileCustomizationBuy(moveAction = null, _reply = null, playerProfile = null) {
         const output = {
