+++ lib/controllers/index.js
@@ -8,6 +8,7 @@ const LocationController = require("./LocationController");
 const AccountController = require("./AccountController");
 const WeblauncherController = require("./WeblauncherController");
 const RaidController = require("./RaidController");
+const ItemController = require("./ItemController");
 
 module.exports = {
     ...AccountController,
@@ -19,5 +20,6 @@ module.exports = {
     ...MenuController,
     ...TradingController,
     ...WeblauncherController,
-    ...RaidController
+    ...RaidController,
+    ...ItemController
 };
